Fixed type error.
R=vsm@google.com

Review URL: https://codereview.chromium.org//34103002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@28974 260f80e4-7a28-3924-810f-c04153c831b5

+++ tests/html/canvasrenderingcontext2d_test.dart
@@ -689,7 +689,7 @@ main() {
       // Draw a blue box.
       context.fillText('█', x, y);
 
-      var width = context.measureText('█').width;
+      var width = context.measureText('█').width.toInt();
 
       checkPixel(readPixel(x, y), [0, 0, 255, 255]);
       checkPixel(readPixel(x + 10, y), [0, 0, 255, 255]);
@@ -710,7 +710,7 @@ main() {
       // Draw a blue box with null maxWidth.
       context.fillText('█', x, y, null);
 
-      var width = context.measureText('█').width;
+      var width = context.measureText('█').width.toInt();
 
       checkPixel(readPixel(x, y), [0, 0, 255, 255]);
       checkPixel(readPixel(x + 10, y), [0, 0, 255, 255]);
